+++ sigstore/models.py
@@ -415,6 +415,27 @@ class InvalidBundle(Error):
         )
 
 
+class IncompatibleEntry(InvalidBundle):
+    """
+    Raised when the log entry within the `Bundle` has an incompatible KindVersion.
+
+    The common case is a Rekor v2 entry.
+    """
+
+    def diagnostics(self) -> str:
+        """Returns diagnostics for the error."""
+
+        return dedent(
+            f"""\
+        The provided bundle contains a transparency log entry that is incompatible with this version of sigstore-python. Please upgrade your verifying client.
+
+        Additional context:
+
+        {self}
+        """
+        )
+
+
 class Bundle:
     """
     Represents a Sigstore bundle.
@@ -512,6 +533,11 @@ class Bundle:
         if len(tlog_entries) != 1:
             raise InvalidBundle("expected exactly one log entry in bundle")
         tlog_entry = tlog_entries[0]
+
+        if tlog_entry.kind_version.version != "0.0.1":
+            raise IncompatibleEntry(
+                f"Expected log entry version 0.0.1, got {tlog_entry.kind_version.version}"
+            )
 
         # Handling of inclusion promises and proofs varies between bundle
         # format versions: